Refactor the metrics to define their own aggregations
