+++ nucleus/metrics/polygon_metrics.py
@@ -83,6 +83,9 @@ class PolygonMetric(Metric):
         # Main evaluation function that subclasses must override.
         pass
 
+    def aggregate(self, results: List[ScalarResult]) -> ScalarResult:  # type: ignore[override]
+        return ScalarResult.aggregate(results)
+
     def __call__(
         self, annotations: AnnotationList, predictions: PredictionList
     ) -> ScalarResult:
